# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API server for managing Kibaship Operator resources",
    "title": "Kibaship Operator API",
    "termsOfService": "https://github.com/kibamail/kibaship",
    "contact": {
      "name": "API Support",
      "email": "support@kibamail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
...
    "/v1/environments/{uuid}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve an environment by its unique UUID or slug identifier",
        "produces": ["application/json"],
        "tags": ["environments"],
        "summary": "Get environment by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Environment UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Environment details",
            "schema": {
              "$ref": "#/definitions/models.EnvironmentResponse"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Environment not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete an environment by its unique UUID or slug identifier",
        "tags": ["environments"],
        "summary": "Delete environment by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Environment UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Environment deleted successfully"
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Environment not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update an environment by its unique UUID or slug identifier with partial updates",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["environments"],
        "summary": "Update environment by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Environment UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          },
          {
            "description": "Environment update data",
            "name": "environment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.EnvironmentUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated environment details",
            "schema": {
              "$ref": "#/definitions/models.EnvironmentResponse"
            }
          },
          "400": {
            "description": "Validation errors in request data",
            "schema": {
              "$ref": "#/definitions/models.ValidationErrors"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Environment not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/v1/projects/{uuid}/environments": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve all environments for a specific project",
        "produces": ["application/json"],
        "tags": ["environments"],
        "summary": "List environments for a project",
        "parameters": [
          {
            "type": "string",
            "description": "Project UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of environments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.EnvironmentResponse"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Project not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new environment within a project. The environment will be assigned a random 8-character slug.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["environments"],
        "summary": "Create a new environment",
        "parameters": [
          {
            "type": "string",
            "description": "Project UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          },
          {
            "description": "Environment creation data",
            "name": "environment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.EnvironmentCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Environment created successfully",
            "schema": {
              "$ref": "#/definitions/models.EnvironmentResponse"
            }
          },
          "400": {
            "description": "Validation errors in request data",
            "schema": {
              "$ref": "#/definitions/models.ValidationErrors"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Project not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "auth.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Unauthorized"
        },
        "message": {
          "type": "string",
          "example": "Invalid API key"
        }
      }
    },
...
    "models.EnvironmentCreateRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Production environment"
        },
        "name": {
          "type": "string",
          "example": "production"
        },
        "projectUuid": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426614174001"
        },
        "variables": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "models.EnvironmentResponse": {
      "type": "object",
      "properties": {
        "applicationCount": {
          "type": "integer",
          "example": 5
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "description": {
          "type": "string",
          "example": "Production environment"
        },
        "name": {
          "type": "string",
          "example": "production"
        },
        "projectSlug": {
          "type": "string",
          "example": "xyz789ab"
        },
        "projectUuid": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426614174001"
        },
        "slug": {
          "type": "string",
          "example": "abc123de"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "uuid": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426614174000"
        },
        "variables": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "models.EnvironmentUpdateRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "Updated production environment"
        },
        "variables": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "models.ValidationError": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "models.ValidationErrors": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.ValidationError"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}